Converted project to typescript
