+++ package.json
@@ -23,6 +23,10 @@
     "inquirer": "^9.1.4"
   },
   "devDependencies": {
-    "eslint": "^8.29.0"
+    "@types/inquirer": "^9.0.3",
+    "@types/node": "^18.11.12",
+    "@typescript-eslint/eslint-plugin": "^5.46.0",
+    "eslint": "^8.29.0",
+    "typescript": "^4.9.4"
   }
 }
