+++ src/lib/BSD-2-Clause-Simplified-License.ts
@@ -1,6 +1,6 @@
-export const BSD3License = (year, author) => {
+export const BSD2License = (year: number, author: string): string => {
     return `
-BSD 3-Clause License
+BSD 2-Clause License
 
 Copyright (c) ${year}, ${author}
 All rights reserved.
@@ -14,10 +14,6 @@ modification, are permitted provided that the following conditions are met:
 2. Redistributions in binary form must reproduce the above copyright notice,
    this list of conditions and the following disclaimer in the documentation
    and/or other materials provided with the distribution.
-
-3. Neither the name of the copyright holder nor the names of its
-   contributors may be used to endorse or promote products derived from
-   this software without specific prior written permission.
 
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
