+++ src/lib/Boost-Software-License-1.ts
@@ -1,4 +1,4 @@
-export const BoostLicense = () => {
+export const BoostLicense = (): string => {
     return `
 Boost Software License - Version 1.0 - August 17th, 2003
