+++ src/types.ts
@@ -1,0 +1,11 @@
+export interface InputData {
+    title: string;
+    license: string;
+    description: string;
+    install: string;
+    usage: string;
+    contribution: string;
+    test: string;
+    email: string;
+    user: string;
+}
