+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+    "compilerOptions": {
+      "module": "ES6",
+      "esModuleInterop": true,
+      "target": "es6",
+      "moduleResolution": "node"
+    },
+    "include": ["./src/**/*"]
+}
